changed the wordings for the footer.
Let's boot up and dig into our heels! Start coding, move up to be a power coder at your convenience. The more we code, the more we learn and the more we grow :). Let us pool our brains together to throw up a an elegant and efficient website for our Alma Mater - Toc H Institute of Science and Technology.